split out settings view from settings
