+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 members = [
 	"spel-katalog",
+	"spel-katalog-daemon",
+	"spel-katalog-install",
 	"spel-katalog/cli",
 	"spel-katalog/common",
 	"spel-katalog/formats",
@@ -9,6 +11,7 @@ members = [
 	"spel-katalog/info",
 	"spel-katalog/installer",
 	"spel-katalog/ipc",
+	"spel-katalog/lazy",
 	"spel-katalog/native",
 	"spel-katalog/profiler",
 	"spel-katalog/run",
@@ -16,6 +19,7 @@ members = [
 	"spel-katalog/settings/derive",
 	"spel-katalog/settings/derive-lib",
 	"spel-katalog/settings/traits",
+	"spel-katalog/settings/view",
 	"spel-katalog/sink",
 	"spel-katalog/terminal",
 	"spel-katalog/widget",
@@ -154,3 +158,6 @@ vte = "0.15.0"
 whoami = { version = "2.1.2", default-features = false }
 xdg = "3.0.0"
 yaml-rust2 = "0.11.0"
+spel-katalog-install = { path = "spel-katalog-install" }
+spel-katalog-lazy = { path = "spel-katalog/lazy" }
+spel-katalog-settings-view = { path = "spel-katalog/settings/view" }
