+++ spel-katalog-daemon/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "spel-katalog-daemon"
+version = "0.1.0"
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
