+++ spel-katalog-daemon/src/main.rs
@@ -1,0 +1,5 @@
+//! Application daemon.
+
+fn main() {
+    println!("Hello, world!");
+}
