+++ spel-katalog/formats/Cargo.toml
@@ -11,13 +11,13 @@ chrono = { workspace = true, features = ["clock", "std"] }
 derive_more = { workspace = true, features = ["is_variant", "display", "from", "into", "deref", "deref_mut", "as_ref"] }
 image.workspace = true
 indexmap.workspace = true
+log.workspace = true
 rustc-hash.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_bytes.workspace = true
 strum = { workspace = true, features = ["derive"] }
 thiserror.workspace = true
 uuid = { workspace = true, features = ["serde"] }
-log.workspace = true
 yaml-rust2.workspace = true
 
 [lints]
