+++ spel-katalog/installer/Cargo.toml
@@ -23,6 +23,7 @@ spel-katalog-common.workspace = true
 spel-katalog-formats.workspace = true
 spel-katalog-native.workspace = true
 spel-katalog-settings.workspace = true
+spel-katalog-settings-view.workspace = true
 spel-katalog-widget.workspace = true
 tap.workspace = true
 toml.workspace = true
