+++ spel-katalog/settings/Cargo.toml
@@ -8,19 +8,14 @@ license.workspace = true
 [dependencies]
 clap = { workspace = true, features = ["derive"] }
 derive_more = { workspace = true, features = ["from", "is_variant", "display", "into_iterator", "as_ref", "deref", "into", "deref_mut"] }
-iced_core.workspace = true
-iced_renderer.workspace = true
-iced_runtime.workspace = true
-iced_widget.workspace = true
 log.workspace = true
+rustix.workspace = true
 serde = { workspace = true, features = ["derive"] }
 smol.workspace = true
-spel-katalog-common.workspace = true
-spel-katalog-widget.workspace = true
 spel-katalog-settings-traits.workspace = true
+spel-katalog-lazy.workspace = true
 tap.workspace = true
 toml.workspace = true
-rustix.workspace = true
 whoami = { workspace = true }
 xdg.workspace = true
