+++ spel-katalog/src/main.rs
@@ -1,13 +1,8 @@
-use ::std::{
-    collections::HashMap,
-    io::{IsTerminal, Read},
-};
+use ::std::io::{IsTerminal, Read};
 
-use ::color_eyre::{Section, eyre::eyre};
 use ::mimalloc::MiMalloc;
 use ::spel_katalog::run as run_app;
 use ::spel_katalog_cli::{Cli, InstallGame, Subcmd, SubcmdCallbacks};
-use ::spel_katalog_formats::{Bind, InstallerConfig};
 use ::spel_katalog_sink::SinkBuilder;
 
 #[global_allocator]
@@ -43,58 +38,9 @@ fn run(cli: ::spel_katalog_cli::Run) -> ::color_eyre::Result<()> {
     Ok(())
 }
 
-fn install_game(
-    InstallGame {
-        game,
-        thumbnail,
-        hidden,
-        no_move,
-        exe,
-        installer_dir,
-        dummy,
-    }: InstallGame,
-) -> ::color_eyre::Result<()> {
+fn install_game(install_game: InstallGame) -> ::color_eyre::Result<()> {
     init_log(None);
-    let base_dirs = ::xdg::BaseDirectories::with_prefix("spel-katalog");
-    let drives = installer_dir
-        .as_ref()
-        .map(|d| HashMap::from_iter([('i', d.clone())]))
-        .unwrap_or_default();
-    let ro_bind = installer_dir
-        .map(|d| Vec::from_iter([Bind::mirrored(d)]))
-        .unwrap_or_default();
-    let game_dir = game
-        .canonicalize()
-        .map_err(|err| eyre!(err).note(format!("is {game:?} a valid path?")))?;
-    let exe = if !dummy {
-        exe
-    } else {
-        Some(game_dir.join("dummy.exe"))
-    };
-
-    if let Err(err) = ::spel_katalog_ipc::send(
-        &base_dirs,
-        ::spel_katalog_ipc::Message::InstallGame(InstallerConfig {
-            game_dir,
-            exe,
-            hidden: Some(hidden),
-            thumbnail: thumbnail
-                .map(|t| {
-                    t.canonicalize()
-                        .map_err(|err| eyre!(err).note(format!("is {t:?} a valid path?")))
-                })
-                .transpose()?,
-            move_game: Some(!no_move),
-            drives,
-            bind: Default::default(),
-            ro_bind,
-            env: Default::default(),
-        }),
-    ) {
-        Err(eyre!(err).note("is the application open?"))
-    } else {
-        Ok(())
-    }
+    ::spel_katalog_install::install_game(install_game)
 }
 
 fn main() -> ::color_eyre::Result<()> {
